Move custom field and form classes to utils dir
