+++ kqueen_ui/utils/fields.py
@@ -1,0 +1,87 @@
+from flask_wtf.file import FileField as WTFileField
+from werkzeug.datastructures import FileStorage
+from wtforms import (
+    IntegerField as WTIntegerField,
+    PasswordField as WTPasswordField,
+    SelectField as WTSelectField,
+    StringField as WTStringField,
+    TextAreaField as WTTextAreaField
+)
+from wtforms.fields.html5 import EmailField as WTEmailField
+
+import json
+import logging
+import yaml
+
+logger = logging.getLogger(__name__)
+
+
+# MIXINS
+
+class SelectableMixin:
+    switchtag = None
+    jsvalidators = {}
+
+    def __init__(self, *args, **kwargs):
+        self.switchtag = kwargs.pop('switchtag', None)
+        self.jsvalidators = kwargs.pop('jsvalidators', None)
+        super(SelectableMixin, self).__init__(*args, **kwargs)
+
+
+# FIELDS
+
+class EmailField(SelectableMixin, WTEmailField):
+    pass
+
+
+class FileField(SelectableMixin, WTFileField):
+    pass
+
+
+class IntegerField(SelectableMixin, WTIntegerField):
+    pass
+
+
+class JsonFileField(FileField):
+
+    def process_formdata(self, valuelist):
+        super(JsonFileField, self).process_formdata(valuelist)
+        if self.data and isinstance(self.data, FileStorage):
+            try:
+                data = self.data.read()
+                self.data = json.loads(data.decode('utf-8'))
+            except Exception as e:
+                self.data = {}
+                logger.error('Could not load JSON file: {}'.format(repr(e)))
+
+
+class PasswordField(SelectableMixin, WTPasswordField):
+    pass
+
+
+class SelectField(SelectableMixin, WTSelectField):
+    switch = False
+
+    def __init__(self, *args, **kwargs):
+        self.switch = kwargs.pop('switch', False)
+        super(SelectField, self).__init__(*args, **kwargs)
+
+
+class StringField(SelectableMixin, WTStringField):
+    pass
+
+
+class TextAreaField(SelectableMixin, WTTextAreaField):
+    pass
+
+
+class YamlFileField(FileField):
+
+    def process_formdata(self, valuelist):
+        super(YamlFileField, self).process_formdata(valuelist)
+        if self.data and isinstance(self.data, FileStorage):
+            try:
+                self.data = yaml.load(self.data.stream)
+            except Exception as e:
+                self.data = {}
+                logger.error('Could not load YAML file: {}'.format(repr(e)))
